+++ solarforecastarbiter/reference_forecasts/main.py
@@ -14,8 +14,7 @@ import pandas as pd
 
 from solarforecastarbiter import datamodel, pvmodel
 from solarforecastarbiter.io.fetch import nwp as fetch_nwp
-from solarforecastarbiter.io import utils as io_utils
-from solarforecastarbiter.reference_forecasts import persistence, models
+from solarforecastarbiter.reference_forecasts import persistence, models, utils
 
 
 logger = logging.getLogger(__name__)
@@ -173,13 +172,14 @@ def run_persistence(session, observation, forecast, run_time, issue_time,
     ValueError
         If average observations are used to make instantaneous forecast.
     """
-    forecast_start, forecast_end = get_forecast_start_end(forecast, issue_time)
-    intraday = _is_intraday(forecast)
+    forecast_start, forecast_end = utils.get_forecast_start_end(
+        forecast, issue_time)
+    intraday = utils._is_intraday(forecast)
     if not intraday:
         # raise ValueError if not intraday and not midnight to midnight
-        _check_midnight_to_midnight(forecast_start, forecast_end)
+        utils._check_midnight_to_midnight(forecast_start, forecast_end)
 
-    data_start, data_end = get_data_start_end(
+    data_start, data_end = utils.get_data_start_end(
         observation, forecast, run_time)
 
     def load_data(observation, data_start, data_end):
@@ -206,139 +206,6 @@ def run_persistence(session, observation, forecast, run_time, issue_time,
             'index=True not supported for forecasts with run_length >= 1day')
 
     return fx
-
-
-def get_forecast_start_end(forecast, issue_time):
-    """
-    Get absolute forecast start from *forecast* object parameters and
-    absolute *issue_time*.
-
-    Parameters
-    ----------
-    forecast : datamodel.Forecast
-    issue_time : pd.Timestamp
-
-    Returns
-    -------
-    forecast_start : pd.Timestamp
-        Start time of forecast issued at issue_time
-    forecast_end : pd.Timestamp
-        End time of forecast issued at issue_time
-
-    Raises
-    ------
-    ValueError if forecast and issue_time are incompatible
-    """
-    first_issue_time = pd.Timestamp.combine(issue_time.floor('1D'),
-                                            forecast.issue_time_of_day)
-    issue_times = pd.date_range(start=first_issue_time,
-                                end=first_issue_time+pd.Timedelta('1d'),
-                                freq=forecast.run_length)
-    if issue_time not in issue_times:
-        raise ValueError(
-            ('Incompatible forecast.issue_time_of_day %s, '
-             'forecast.run_length %s, and issue_time %s') % (
-             forecast.issue_time_of_day, forecast.run_length, issue_time))
-    forecast_start = issue_time + forecast.lead_time_to_start
-    forecast_end = forecast_start + forecast.run_length
-    return io_utils.adjust_start_end_for_interval_label(
-        forecast.interval_label, forecast_start, forecast_end)
-
-
-def _is_intraday(forecast):
-    """Is the forecast intraday?"""
-    # intra day persistence and "day ahead" persistence require
-    # fairly different parameters.
-    # is this a sufficiently robust way to distinguish?
-    return forecast.run_length < pd.Timedelta('1d')
-
-
-def _check_midnight_to_midnight(forecast_start, forecast_end):
-    if (forecast_start.round('1d') != forecast_start or
-            forecast_end - forecast_start > pd.Timedelta('1d')):
-        raise ValueError(
-            'Day ahead persistence requires midnight to midnight periods')
-
-
-def _intraday_start_end(observation, forecast, run_time):
-    # time window over which observation data will be used to create
-    # persistence forecast.
-    if (observation.interval_length > forecast.run_length or
-            observation.interval_length > pd.Timedelta('1h')):
-        raise ValueError(
-            'Intraday persistence requires observation.interval_length '
-            '<= forecast.run_length and observation.interval_length <= 1h')
-    # no longer than 1 hour
-    window = min(forecast.run_length, pd.Timedelta('1hr'))
-    data_end = run_time
-    data_start = data_end - window
-    return data_start, data_end
-
-
-def _dayahead_start_end(run_time):
-    # day ahead persistence: tomorrow's forecast is equal to yesterday's
-    # observations. So, forecast always uses obs > 24 hr old at each valid
-    # time. Smarter approach might be to use today's observations up
-    # until issue_time, and use yesterday's observations for issue_time
-    # until end of day. So, forecast *never* uses obs > 24 hr old at each
-    # valid time. Arguably too much for a reference forecast.
-    data_end = run_time.floor('1d')
-    data_start = data_end - pd.Timedelta('1d')
-    return data_start, data_end
-
-
-def _adjust_for_instant_obs(data_start, data_end, observation, forecast):
-    # instantaneous observations require care.
-    # persistence models return forecasts with same closure as obs
-    if 'instant' in forecast.interval_label:
-        if forecast.interval_length != observation.interval_length:
-            raise ValueError('Instantaneous forecast requires instantaneous '
-                             'observation with identical interval length.')
-        else:
-            data_end -= pd.Timedelta('1s')
-    elif forecast.interval_label == 'beginning':
-        data_end -= pd.Timedelta('1s')
-    else:
-        data_start += pd.Timedelta('1s')
-    return data_start, data_end
-
-
-def get_data_start_end(observation, forecast, run_time):
-    """
-    Determine the data start and data end times for a persistence
-    forecast.
-
-    Returns
-    -------
-    data_start : pd.Timestamp
-    data_end : pd.Timestamp
-    """
-    if _is_intraday(forecast):
-        data_start, data_end = _intraday_start_end(observation, forecast,
-                                                   run_time)
-    else:
-        data_start, data_end = _dayahead_start_end(run_time)
-
-    # to ensure that each observation data point contributes to the correct
-    # forecast, the data_end and data_start values may need to be nudged
-    if 'instant' in observation.interval_label:
-        data_start, data_end = _adjust_for_instant_obs(data_start, data_end,
-                                                       observation, forecast)
-    else:
-        if 'instant' in forecast.interval_label:
-            raise ValueError('Instantaneous forecast cannot be made from '
-                             'interval average observations')
-    return data_start, data_end
-
-
-def get_init_time(run_time, fetch_metadata):
-    """Determine the most recent init time for which all forecast data is
-    available."""
-    run_finish = (pd.Timedelta(fetch_metadata['delay_to_first_forecast']) +
-                  pd.Timedelta(fetch_metadata['avg_max_run_length']))
-    freq = fetch_metadata['update_freq']
-    init_time = (run_time - run_finish).floor(freq=freq)
-    return init_time
 
 
 def process_forecast_groups(forecasts, session, run_time, issue_time,
@@ -369,8 +236,8 @@ def process_forecast_groups(forecasts, session, run_time, issue_time,
                     'not generating forecasts for group',
                     key_fx.name)
             continue
-        init_time = get_init_time(run_time, fetch_metadata)
-        forecast_start, forecast_end = get_forecast_start_end(
+        init_time = utils.get_init_time(run_time, fetch_metadata)
+        forecast_start, forecast_end = utils.get_forecast_start_end(
             key_fx, issue_time)
         model = getattr(models, extra_params['model'])
         # for testing
