+++ solarforecastarbiter/reference_forecasts/tests/test_main.py
@@ -64,252 +64,6 @@ def test_run_nwp(model, load_forecast_return_value, site_powerplant_site_type,
     check_out(out, out_forecast_exp, site_type)
 
 
-def test_get_data_start_end_labels_1h_window_limit(site_metadata):
-    observation = default_observation(
-        site_metadata,
-        interval_length=pd.Timedelta('5min'), interval_label='beginning')
-    forecast = default_forecast(
-        site_metadata,
-        run_length=pd.Timedelta('12h'),  # test 1 hr limit on window
-        interval_label='beginning')
-    # ensure data no later than run time
-    run_time = pd.Timestamp('20190422T1945Z')
-    data_start, data_end = main.get_data_start_end(observation, forecast,
-                                                   run_time)
-    assert data_start == pd.Timestamp('20190422T1845Z')
-    assert data_end == pd.Timestamp('20190422T1945Z')
-
-
-def test_get_data_start_end_labels_subhourly_window_limit(site_metadata):
-    observation = default_observation(
-        site_metadata,
-        interval_length=pd.Timedelta('5min'), interval_label='beginning')
-    forecast = default_forecast(
-        site_metadata,
-        run_length=pd.Timedelta('5min'),  # test subhourly limit on window
-        interval_label='beginning')
-    run_time = pd.Timestamp('20190422T1945Z')
-    data_start, data_end = main.get_data_start_end(observation, forecast,
-                                                   run_time)
-    assert data_start == pd.Timestamp('20190422T1940Z')
-    assert data_end == pd.Timestamp('20190422T1945Z')
-
-
-def test_get_data_start_end_labels_obs_longer_than_fx_run(site_metadata):
-    observation = default_observation(
-        site_metadata,
-        interval_length=pd.Timedelta('15min'))
-    forecast = default_forecast(
-        site_metadata,
-        run_length=pd.Timedelta('5min'))
-    run_time = pd.Timestamp('20190422T1945Z')
-    # obs interval cannot be longer than forecast interval
-    with pytest.raises(ValueError) as excinfo:
-        main.get_data_start_end(observation, forecast, run_time)
-    assert ("observation.interval_length <= forecast.run_length" in
-            str(excinfo.value))
-
-
-def test_get_data_start_end_labels_obs_longer_than_1h(site_metadata):
-    observation = default_observation(
-        site_metadata,
-        interval_length=pd.Timedelta('2h'))
-    forecast = default_forecast(
-        site_metadata,
-        run_length=pd.Timedelta('5min'))
-    run_time = pd.Timestamp('20190422T1945Z')
-    # obs interval cannot be longer than 1 hr
-    with pytest.raises(ValueError) as excinfo:
-        main.get_data_start_end(observation, forecast, run_time)
-    assert 'observation.interval_length <= 1h' in str(excinfo.value)
-
-
-def test_get_data_start_end_labels_obs_longer_than_1h_day_ahead(site_metadata):
-    observation = default_observation(
-        site_metadata,
-        interval_length=pd.Timedelta('2h'), interval_label='beginning')
-    forecast = default_forecast(
-        site_metadata,
-        issue_time_of_day=dt.time(hour=5),
-        lead_time_to_start=pd.Timedelta('1h'),
-        interval_length=pd.Timedelta('1h'),
-        run_length=pd.Timedelta('1d'),  # day ahead
-        interval_label='beginning')
-    run_time = pd.Timestamp('20190422T1945Z')
-    # day ahead doesn't care about obs interval length
-    data_start, data_end = main.get_data_start_end(observation, forecast,
-                                                   run_time)
-    assert data_start == pd.Timestamp('20190421T0000Z')
-    assert data_end == pd.Timestamp('20190422T0000Z')
-
-
-def test_get_data_start_end_labels_obs_fx_instant_mismatch(site_metadata):
-    observation = default_observation(
-        site_metadata,
-        interval_length=pd.Timedelta('5min'), interval_label='instant')
-    forecast = default_forecast(
-        site_metadata,
-        issue_time_of_day=dt.time(hour=5),
-        lead_time_to_start=pd.Timedelta('1h'),
-        interval_length=pd.Timedelta('1h'),  # interval_length must be equal
-        run_length=pd.Timedelta('1d'),
-        interval_label='instant')            # if interval_label also instant
-    run_time = pd.Timestamp('20190422T1945Z')
-    with pytest.raises(ValueError) as excinfo:
-        main.get_data_start_end(observation, forecast, run_time)
-    assert 'with identical interval length' in str(excinfo.value)
-
-
-def test_get_data_start_end_labels_obs_fx_instant(site_metadata):
-    observation = default_observation(
-        site_metadata,
-        interval_length=pd.Timedelta('5min'), interval_label='instant')
-    forecast = default_forecast(
-        site_metadata,
-        issue_time_of_day=dt.time(hour=5),
-        lead_time_to_start=pd.Timedelta('1h'),
-        interval_length=pd.Timedelta('5min'),  # interval_length must be equal
-        run_length=pd.Timedelta('1d'),
-        interval_label='instant')              # if interval_label also instant
-    run_time = pd.Timestamp('20190422T1945Z')
-    data_start, data_end = main.get_data_start_end(observation, forecast,
-                                                   run_time)
-    assert data_start == pd.Timestamp('20190421T0000Z')
-    assert data_end == pd.Timestamp('20190421T235959Z')
-
-
-def test_get_data_start_end_labels_obs_instant_fx_avg(site_metadata):
-    observation = default_observation(
-        site_metadata,
-        interval_length=pd.Timedelta('5min'), interval_label='instant')
-    forecast = default_forecast(
-        site_metadata,
-        issue_time_of_day=dt.time(hour=5),
-        lead_time_to_start=pd.Timedelta('1h'),
-        interval_length=pd.Timedelta('5min'),
-        run_length=pd.Timedelta('1d'),
-        interval_label='beginning')
-    run_time = pd.Timestamp('20190422T1945Z')
-    data_start, data_end = main.get_data_start_end(observation, forecast,
-                                                   run_time)
-    assert data_start == pd.Timestamp('20190421T0000Z')
-    assert data_end == pd.Timestamp('20190421T235959Z')
-
-
-def test_get_data_start_end_labels_obs_instant_fx_avg_ending(site_metadata):
-    observation = default_observation(
-        site_metadata,
-        interval_length=pd.Timedelta('5min'), interval_label='instant')
-    forecast = default_forecast(
-        site_metadata,
-        issue_time_of_day=dt.time(hour=5),
-        lead_time_to_start=pd.Timedelta('1h'),
-        interval_length=pd.Timedelta('5min'),
-        run_length=pd.Timedelta('1d'),
-        interval_label='ending')
-    run_time = pd.Timestamp('20190422T1945Z')
-    data_start, data_end = main.get_data_start_end(observation, forecast,
-                                                   run_time)
-    assert data_start == pd.Timestamp('20190421T000001Z')
-    assert data_end == pd.Timestamp('20190422T0000Z')
-
-
-def test_get_data_start_end_labels_obs_instant_fx_avg_intraday(site_metadata):
-    run_time = pd.Timestamp('20190422T1945Z')
-    observation = default_observation(
-        site_metadata,
-        interval_length=pd.Timedelta('5min'), interval_label='instant')
-    forecast = default_forecast(
-        site_metadata,
-        issue_time_of_day=dt.time(hour=5),
-        lead_time_to_start=pd.Timedelta('1h'),
-        interval_length=pd.Timedelta('5min'),
-        run_length=pd.Timedelta('15min'),
-        interval_label='ending')
-    data_start, data_end = main.get_data_start_end(observation, forecast,
-                                                   run_time)
-    assert data_start == pd.Timestamp('20190422T193001Z')
-    assert data_end == pd.Timestamp('20190422T1945Z')
-
-
-def test_get_data_start_end_labels_obs_avg_fx_instant(site_metadata):
-    run_time = pd.Timestamp('20190422T1945Z')
-    observation = default_observation(
-        site_metadata,
-        interval_length=pd.Timedelta('5min'), interval_label='ending')
-    forecast = default_forecast(
-        site_metadata,
-        issue_time_of_day=dt.time(hour=5),
-        lead_time_to_start=pd.Timedelta('1h'),
-        interval_length=pd.Timedelta('5min'),
-        run_length=pd.Timedelta('1d'),
-        interval_label='instant')
-    with pytest.raises(ValueError) as excinfo:
-        main.get_data_start_end(observation, forecast, run_time)
-    assert 'made from interval average obs' in str(excinfo.value)
-
-
-@pytest.fixture
-def forecast_hr_begin(site_metadata):
-    return default_forecast(
-        site_metadata,
-        issue_time_of_day=dt.time(hour=5),
-        lead_time_to_start=pd.Timedelta('1h'),
-        interval_length=pd.Timedelta('1h'),
-        run_length=pd.Timedelta('1h'),
-        interval_label='beginning')
-
-
-def test_get_forecast_start_end_time_same_as_issue_time(forecast_hr_begin):
-    # time is same as issue time of day
-    issue_time = pd.Timestamp('20190422T0500')
-    fx_start, fx_end = main.get_forecast_start_end(
-        forecast_hr_begin, issue_time)
-    assert fx_start == pd.Timestamp('20190422T0600')
-    assert fx_end == pd.Timestamp('20190422T0700')
-
-
-def test_get_forecast_start_end_time_same_as_issue_time_n_x_run(
-        forecast_hr_begin):
-    # time is same as issue time of day + n * run_length
-    issue_time = pd.Timestamp('20190422T1200')
-    fx_start, fx_end = main.get_forecast_start_end(
-        forecast_hr_begin, issue_time)
-    assert fx_start == pd.Timestamp('20190422T1300')
-    assert fx_end == pd.Timestamp('20190422T1400')
-
-
-def test_get_forecast_start_end_time_before_issue_time(forecast_hr_begin):
-    # time is before issue time of day but otherwise valid
-    issue_time = pd.Timestamp('20190422T0400')
-    with pytest.raises(ValueError):
-        fx_start, fx_end = main.get_forecast_start_end(
-            forecast_hr_begin, issue_time)
-
-
-def test_get_forecast_start_end_time_invalid(forecast_hr_begin):
-    # time is invalid
-    issue_time = pd.Timestamp('20190423T0505')
-    with pytest.raises(ValueError):
-        fx_start, fx_end = main.get_forecast_start_end(
-            forecast_hr_begin, issue_time)
-
-
-def test_get_forecast_start_end_time_instant(site_metadata):
-    # instant
-    forecast = default_forecast(
-        site_metadata,
-        issue_time_of_day=dt.time(hour=5),
-        lead_time_to_start=pd.Timedelta('1h'),
-        interval_length=pd.Timedelta('5min'),
-        run_length=pd.Timedelta('1h'),
-        interval_label='instant')
-    issue_time = pd.Timestamp('20190422T0500')
-    fx_start, fx_end = main.get_forecast_start_end(forecast, issue_time)
-    assert fx_start == pd.Timestamp('20190422T0600')
-    assert fx_end == pd.Timestamp('20190422T065959')
-
-
 @pytest.fixture
 def obs_5min_begin(site_metadata):
     observation = default_observation(
